+++ src/messages/en.json
@@ -559,7 +559,7 @@
       },
       "username": {
         "label": "Username",
-        "placeholder": "Please add a title",
+        "placeholder": "Please add a username",
         "description": "Username can only contain letters, numbers, underscores, and periods (for SNS names like .sol)",
         "validation": {
           "min_length": "Username must be at least 3 characters long",
